ì코드리뷰로 인한  함수 위치 수정
